tree: Add `PhoshStatusIconsBox`
This box is roughly similar to a `GtkBox`, but instead of trying to
squeeze all children within given space, it only allocates for available
width and gives null allocation for the remaining. As the name says, its
children can be only `PhoshStatusIcon` or `PhoshRevealer` with a
`PhoshStatusIcon` as child.

Signed-off-by: Arun Mani J <j.arunmani@proton.me>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1743>