lockscreen-manager: Add a brightness gesture
Two finger swipe left to right increases brightness. The new
brightness is calculated as offset to the current brightness
based on the swipe distance.

This means that if something else sets brightness in the meantime it
will be overwritten. If we'd just add the difference between two events
this might be too small to make an actual difference thus increasing the
brightness too slowly.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/814
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1831>